管理者権限の追加/編集・削除の制限

+++ app/Http/Controllers/UserController.php
@@ -85,6 +85,7 @@ class UserController extends Controller
   // 更新用フォームへ移動
   public function edit(User $user)
   {
+    $this->authorize('edit', $user);
     return view('users.edit', ['user' => $user]);
   }
 
@@ -99,6 +100,7 @@ class UserController extends Controller
   // 完了後、更新したユーザーのページへ移動
   public function update(Request $request, User $user)
   {
+    $this->authorize('edit', $user);
     $user->name = $request->name;
     $user->save();
     return redirect('users/' . $user->id);
@@ -113,6 +115,7 @@ class UserController extends Controller
   // 一件削除
   public function destroy(User $user)
   {
+    $this->authorize('edit', $user);
     $user->delete();
     return redirect('users');
   }
